clientStreams, rationalising the client side logic to allow plugin output streams, and providing a simple interface to the store. See the programmers guide for details

+++ src/stmem.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: stmem.cc,v 1.70 2001/10/24 08:19:08 hno Exp $
+ * $Id: stmem.cc,v 1.71 2002/09/15 05:41:57 robertc Exp $
  *
  * DEBUG: section 19    Store Memory Primitives
  * AUTHOR: Harvest Derived
@@ -133,9 +133,10 @@ stmemCopy(const mem_hdr * mem, off_t offset, char *buf, size_t size)
     char *ptr_to_buf = NULL;
     int bytes_from_this_packet = 0;
     int bytes_into_this_packet = 0;
-    debug(19, 6) ("memCopy: offset %ld: size %d\n", (long int) offset, (int) size);
+    debug(19, 6) ("memCopy: offset %ld: size %u\n", (long int) offset, size);
     if (p == NULL)
 	return 0;
+    /* RC: the next assert is useless */
     assert(size > 0);
     /* Seek our way into store */
     while ((t_off + p->len) < offset) {
